Ajout selection date page map

+++ src/main/java/fr/amu/iut/sismicviewer/CSV/SeismeDataManager.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 
 public class SeismeDataManager {
 
-    public ArrayList<Seisme> getAnneeFromTo(ArrayList<Seisme> data, double from, double to){
+    public ArrayList<Seisme> getAnneeFromTo(ArrayList<Seisme> data, int from, int to){
         ArrayList<Seisme> seismeTries = new ArrayList<Seisme>();
         for(Seisme seisme : data){
             if(seisme.getAnnee() >= from && seisme.getAnnee() <= to){
