+++ src/main/java/fr/amu/iut/sismicviewer/scenes/dashboard/DashboardControl.java
@@ -92,7 +92,7 @@ public class DashboardControl implements Initializable{
             CSVManager csvManager = new CSVManager();
             SeismeDataManager seismeDataManager = new SeismeDataManager();
             csvManager.loadCsv(SismicViewerApp.getCsvFile());
-            ArrayList<Seisme> dataAnnee = seismeDataManager.getAnneeFromTo(CSVManager.getListeSeisme(), mainRangeSlider.getLowValue(), mainRangeSlider.getHighValue());
+            ArrayList<Seisme> dataAnnee = seismeDataManager.getAnneeFromTo(CSVManager.getListeSeisme(), (int) mainRangeSlider.getLowValue(),  (int) mainRangeSlider.getHighValue());
             mainMapLayer.updateLayer(dataAnnee);
         });
